10.11.22 - Refatoracoes

+++ MinhasFinancas.Domain/Handlers/Financas/MovimentoFinanceiroHandler.cs
@@ -24,7 +24,7 @@ namespace MinhasFinancas.Domain.Handlers.Financas
         {
             if (request is null) return default;
 
-            var entity = new MovimentoFinanceiro(Guid.NewGuid(), request.Valor, request.Titulo, request.Data, (int)request.Tipo, request.ClienteId);
+            var entity = new MovimentoFinanceiro(Guid.NewGuid(), request.Valor, request.Titulo, request.Data, request.Tipo, request.ClienteId);
 
             var result = await _financasRepositorio.GravarMovimentoFinanceiro(entity).ConfigureAwait(false);
